--- a/bestla/bestla/bestla_parallel.h
+++ b/bestla/bestla/bestla_parallel.h
@@ -600,10 +600,14 @@ class OMPThreading : public IThreading {
  public:
   explicit OMPThreading(int nthreads) : IThreading(nthreads) { omp_set_num_threads(nthreads); }
   void parallel_for(const thread_func& func) const override {
+    if (mThreadNum > 1) {
 #pragma omp parallel
-    {
-      int tidx = omp_get_thread_num();
-      func(tidx);
+      {
+        int tidx = omp_get_thread_num();
+        func(tidx);
+      }
+    } else {
+      func(0);
     }
   }
   virtual void set_threads(int nthreads) override {