--- a/bestla/bestla/kernel_avx512f.h
+++ b/bestla/bestla/kernel_avx512f.h
@@ -2225,42 +2225,66 @@ static inline BTLA_CODE layernorm(const float* srcptr, const float* scaleptr, co
   auto vms = _mm512_set1_ps(inv_meansq);
   h = 0;
   if constexpr (simplified) {
-    for (; h < norm_size16; h += VLen) {
-      auto inp = _mm512_loadu_ps(srcptr + h);
-      auto scale = _mm512_loadu_ps(scaleptr + h);
-      inp = _mm512_mul_ps(inp, vms);
-      inp = _mm512_mul_ps(inp, scale);
-      _mm512_storeu_ps(dstptr + h, inp);
-    }
-    for (; h < norm_size; h++) {
-      dstptr[h] = srcptr[h] * inv_meansq * scaleptr[h];
-    }
-  } else {
-    if (biasptr == nullptr) {
+    if (scaleptr) {
       for (; h < norm_size16; h += VLen) {
         auto inp = _mm512_loadu_ps(srcptr + h);
         auto scale = _mm512_loadu_ps(scaleptr + h);
-        inp = _mm512_sub_ps(inp, vm);
         inp = _mm512_mul_ps(inp, vms);
         inp = _mm512_mul_ps(inp, scale);
         _mm512_storeu_ps(dstptr + h, inp);
       }
       for (; h < norm_size; h++) {
-        dstptr[h] = (srcptr[h] - mean) * inv_meansq * scaleptr[h];
+        dstptr[h] = srcptr[h] * inv_meansq * scaleptr[h];
       }
     } else {
       for (; h < norm_size16; h += VLen) {
         auto inp = _mm512_loadu_ps(srcptr + h);
-        auto scale = _mm512_loadu_ps(scaleptr + h);
-        inp = _mm512_sub_ps(inp, vm);
         inp = _mm512_mul_ps(inp, vms);
-        inp = _mm512_mul_ps(inp, scale);
-        auto bias = _mm512_loadu_ps(biasptr + h);
-        inp = _mm512_add_ps(inp, bias);
         _mm512_storeu_ps(dstptr + h, inp);
       }
       for (; h < norm_size; h++) {
-        dstptr[h] = (srcptr[h] - mean) * inv_meansq * scaleptr[h] + biasptr[h];
+        dstptr[h] = srcptr[h] * inv_meansq;
+      }
+    }
+
+  } else {
+    if (scaleptr) {
+      if (biasptr == nullptr) {
+        for (; h < norm_size16; h += VLen) {
+          auto inp = _mm512_loadu_ps(srcptr + h);
+          auto scale = _mm512_loadu_ps(scaleptr + h);
+          inp = _mm512_sub_ps(inp, vm);
+          inp = _mm512_mul_ps(inp, vms);
+          inp = _mm512_mul_ps(inp, scale);
+          _mm512_storeu_ps(dstptr + h, inp);
+        }
+        for (; h < norm_size; h++) {
+          dstptr[h] = (srcptr[h] - mean) * inv_meansq * scaleptr[h];
+        }
+      } else {
+        for (; h < norm_size16; h += VLen) {
+          auto inp = _mm512_loadu_ps(srcptr + h);
+          auto scale = _mm512_loadu_ps(scaleptr + h);
+          inp = _mm512_sub_ps(inp, vm);
+          inp = _mm512_mul_ps(inp, vms);
+          inp = _mm512_mul_ps(inp, scale);
+          auto bias = _mm512_loadu_ps(biasptr + h);
+          inp = _mm512_add_ps(inp, bias);
+          _mm512_storeu_ps(dstptr + h, inp);
+        }
+        for (; h < norm_size; h++) {
+          dstptr[h] = (srcptr[h] - mean) * inv_meansq * scaleptr[h] + biasptr[h];
+        }
+      }
+    } else {
+      for (; h < norm_size16; h += VLen) {
+        auto inp = _mm512_loadu_ps(srcptr + h);
+        inp = _mm512_sub_ps(inp, vm);
+        inp = _mm512_mul_ps(inp, vms);
+        _mm512_storeu_ps(dstptr + h, inp);
+      }
+      for (; h < norm_size; h++) {
+        dstptr[h] = (srcptr[h] - mean) * inv_meansq;
       }
     }
   }