--- a/bestla/bestla/kernel_avx512f.h
+++ b/bestla/bestla/kernel_avx512f.h
@@ -2195,9 +2195,9 @@ struct padding_trans_interleave_cvt<utils::fp16, utils::bf16, 2> {
 };
 #endif
 
-template <bool simplified>
 static inline BTLA_CODE layernorm(const float* srcptr, const float* scaleptr, const float* biasptr, float epsilon,
-                                  int norm_size, float* dstptr, float* mean_out, float* mean_square_out) {
+                                  int norm_size, float* dstptr, float* mean_out, float* mean_square_out,
+                                  bool simplified) {
   int constexpr VLen = 16;
   int norm_size16 = utils::padto_le(norm_size, VLen);
   int h = 0;
@@ -2215,7 +2215,7 @@ static inline BTLA_CODE layernorm(const float* srcptr, const float* scaleptr, co
     mean_square += srcptr[h] * srcptr[h];
   }
   mean = mean / norm_size;
-  if constexpr (simplified) {
+  if (simplified) {
     mean_square = std::sqrt(mean_square / norm_size + epsilon);
   } else {
     mean_square = std::sqrt(mean_square / norm_size - mean * mean + epsilon);
@@ -2224,7 +2224,7 @@ static inline BTLA_CODE layernorm(const float* srcptr, const float* scaleptr, co
   float inv_meansq = 1.f / mean_square;
   auto vms = _mm512_set1_ps(inv_meansq);
   h = 0;
-  if constexpr (simplified) {
+  if (simplified) {
     if (scaleptr) {
       for (; h < norm_size16; h += VLen) {
         auto inp = _mm512_loadu_ps(srcptr + h);