--- a/bestla/bestla/kernel_ref.h
+++ b/bestla/bestla/kernel_ref.h
@@ -1427,9 +1427,9 @@ static inline BTLA_CODE remove_zeropoint_bias(float* accptr, int ldacc, int row,
   return BTLA_CODE::Success;
 }
 
-template <typename T, bool simplified>
+template <typename T>
 static inline BTLA_CODE layernorm(const T* srcptr, const T* scaleptr, const T* biasptr, T epsilon, int norm_size,
-                                  T* dstptr, T* mean_out, T* mean_square_out) {
+                                  T* dstptr, T* mean_out, T* mean_square_out, bool simplified) {
   T mean = 0;
   T mean_square = 0;
 
@@ -1439,13 +1439,13 @@ static inline BTLA_CODE layernorm(const T* srcptr, const T* scaleptr, const T* b
   }
 
   mean = mean / norm_size;
-  if constexpr (simplified) {
+  if (simplified) {
     mean_square = std::sqrt(mean_square / norm_size + epsilon);
   } else {
     mean_square = std::sqrt(mean_square / norm_size - mean * mean + epsilon);
   }
   float inv_mean_square = 1.f / mean_square;
-  if constexpr (simplified) {
+  if (simplified) {
     if (scaleptr) {
       for (int h = 0; h < norm_size; h++) {
         dstptr[h] = srcptr[h] * inv_mean_square * scaleptr[h];