## Type of Change

Implement AVX2 and AVX512F Layernormalization in BesTLA.
Use BesTLA kernel if src0 is contiguous.
~3x speedup.

Before:
GPTJ: perf_total_per_op_us[                    NORM] =   0.477 ms
LLAMA2: perf_total_per_op_us[                RMS_NORM] =   0.362 ms

After:
GPTJ: perf_total_per_op_us[                    NORM] =   0.055 ms
LLAMA2: perf_total_per_op_us[                RMS_NORM] =   0.117 ms