@luoyu-intel 2024-01-31
> I believe @zhewang1-intc implemented gemm+layernorm fusion before, is it necessary?

Yes, it's for ONNX's definition. It can have scale and bias.